Refactor to give us stateful routes
This way, we don't need to keep track of e.g. "Oh, did I add the user everywhere? What if I need to add more information to pass to other views? Oh nooOoooOooo" and it is all in a state object that we either pass along or clone and append to